ICC IA32 Cleanup.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@6746 6f19259b-4bc3-4df7-8a09-765794883524

+++ EdkCompatibilityPkg/Compatibility/FrameworkHiiToUefiHiiThunk/HiiDatabase.c
@@ -280,7 +280,7 @@ HiiFindHandles (
   }
 
   if (Count > *HandleBufferLength) {
-    *HandleBufferLength = (Count * sizeof (FRAMEWORK_EFI_HII_HANDLE));
+    *HandleBufferLength = (UINT16) (Count * sizeof (FRAMEWORK_EFI_HII_HANDLE));
     return EFI_BUFFER_TOO_SMALL;
   }
 
@@ -299,7 +299,7 @@ HiiFindHandles (
 
   }
 
-  *HandleBufferLength = (Count * sizeof (FRAMEWORK_EFI_HII_HANDLE));
+  *HandleBufferLength = (UINT16) (Count * sizeof (FRAMEWORK_EFI_HII_HANDLE));
   return EFI_SUCCESS;
 }
